# src/main/resources/simulate-assessment-swagger.yaml
{
  "swagger": "2.0",
  "info": {
    "description": "这是一个模型仿真工具与跨域多级模型评估接口文档",
    "version": "v1.0.0",
    "title": "模型仿真工具与跨域多级模型评估接口文档",
    "license": {
      "name": "Authorized by ge tao",
      "url": "http://localhost:8888//model-assessment/swagger-ui.html"
    }
  },
  "host": "localhost:8888",
  "basePath": "/model-assessment",
...
  "paths": {
...
    "/model/createModel": {
      "post": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "模型接入",
        "operationId": "createModelUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "tbModelInfo",
          "description": "tbModelInfo",
          "required": true,
          "schema": {
            "$ref": "#/definitions/TbModelInfo"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«TbModelInfo»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/model/deleteModelInfo": {
      "delete": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "删除模型信息",
        "operationId": "deleteModelsUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "ids",
          "in": "query",
          "description": "ids",
          "required": true,
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          },
          "collectionFormat": "multi"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«TbModelInfo»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/model/findModelUnits": {
      "get": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "查询模型单位",
        "operationId": "findModelUnitsUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«List«string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/model/getFieldFromModel": {
      "get": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "获取模型领域信息",
        "operationId": "getFieldFromModelUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«Collection«string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/model/getModelByType": {
      "get": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "根据模型类别获取模型统计信息",
        "operationId": "getModelCountByServiceTypeUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«List«Map«string,object»»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/model/getModelInfo": {
      "get": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "获取模型信息",
        "operationId": "getModelInfoUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "field",
          "in": "query",
          "description": "模型所属领域",
          "required": false,
          "type": "string"
        },
        {
          "name": "isVisible",
          "in": "query",
          "required": false,
          "type": "boolean"
        },
        {
          "name": "modelName",
          "in": "query",
          "description": "模型名称",
          "required": false,
          "type": "string"
        },
        {
          "name": "modelNameZh",
          "in": "query",
          "required": false,
          "type": "string"
        },
        {
          "name": "pageNo",
          "in": "query",
          "description": "第几页",
          "required": false,
          "type": "string",
          "default": "1"
        },
        {
          "name": "pageSize",
          "in": "query",
          "description": "分页大小",
          "required": false,
          "type": "string",
          "default": "10"
        },
        {
          "name": "serviceType",
          "in": "query",
          "description": "业务流程类型",
          "required": false,
          "type": "string"
        },
        {
          "name": "sortField",
          "in": "query",
          "description": "排序字段",
          "required": false,
          "type": "string"
        },
        {
          "name": "sortOrder",
          "in": "query",
          "description": "排序方式",
          "required": false,
          "type": "string"
        },
        {
          "name": "systemId",
          "in": "query",
          "required": false,
          "type": "integer",
          "format": "int32"
        },
        {
          "name": "unit",
          "in": "query",
          "description": "单位名称",
          "required": false,
          "type": "string"
        },
        {
          "name": "userLevel",
          "in": "query",
          "description": "用户层级",
          "required": false,
          "type": "string"
        },
        {
          "name": "vague",
          "in": "query",
          "description": "模糊查询字段",
          "required": false,
          "type": "string"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«IPage«TbModelInfo»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/model/getModelInfoById": {
      "get": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "根据id获取模型信息",
        "operationId": "getModelInfoByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "modelId",
          "in": "query",
          "description": "模型id",
          "required": true,
          "type": "string"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«TbModelInfo»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/model/updateModelInfo": {
      "put": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "更新模型信息",
        "operationId": "updateModelUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "id",
          "description": "模型id",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "in": "body",
          "name": "tbModelInfo",
          "description": "tbModelInfo",
          "required": true,
          "schema": {
            "$ref": "#/definitions/TbModelInfo"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«TbModelInfo»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/model/updateModelStatus": {
      "put": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "模型启动状态控制",
        "operationId": "updateModelStatusUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "id",
          "in": "query",
          "description": "id",
          "required": true,
          "type": "integer",
          "format": "int64"
        },
        {
          "name": "status",
          "in": "query",
          "description": "status",
          "required": false,
          "type": "boolean"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«TbModelInfo»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/model/updateModelVisible": {
      "put": {
        "tags": [
          "模型管理模块"
        ],
        "summary": "更改模型可见性",
        "operationId": "updateModelVisibleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "id",
          "in": "query",
          "description": "id",
          "required": true,
          "type": "integer",
          "format": "int64"
        },
        {
          "name": "visible",
          "in": "query",
          "description": "visible",
          "required": true,
          "type": "boolean"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/sceneController/deleteSceneInfo": {
      "delete": {
        "tags": [
          "作战场景管理模块"
        ],
        "summary": "删除作战场景信息",
        "operationId": "deleteModelsUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
        {
          "name": "ids",
          "in": "query",
          "description": "ids",
          "required": true,
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          },
          "collectionFormat": "multi"
        }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«SceneInfo»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/shareModel/getModelByType": {
      "get": {
        "tags": [
          "模型共享模块"
        ],
        "summary": "根据模型类别获取模型统计信息",
        "operationId": "getModelCountByServiceTypeUsingGET_1",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommonResult«List«Map«string,object»»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "CommonResult«Collection«string»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«Collection«string»»"
    },
    "CommonResult«IPage«TbModelInfo»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/IPage«TbModelInfo»"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«IPage«TbModelInfo»»"
    },
...
    "CommonResult«List«Map«string,object»»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Map«string,object»"
          }
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«List«Map«string,object»»»"
    },
...
    "CommonResult«List«string»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«List«string»»"
    },
...
    "CommonResult«SceneInfo»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/SceneInfo"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«SceneInfo»"
    },
...
    "CommonResult«TbModelInfo»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/TbModelInfo"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«TbModelInfo»"
    },
...
    "CommonResult«object»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "data": {
          "type": "object"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "CommonResult«object»"
    },
...
    "IPage«TbModelInfo»": {
      "type": "object",
      "properties": {
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TbModelInfo"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "IPage«TbModelInfo»"
    },
...
    "Map«string,object»": {
      "type": "object",
      "title": "Map«string,object»",
      "additionalProperties": {
        "type": "object"
      }
    },
...
    "SceneInfo": {
      "type": "object",
      "properties": {
        "describe": {
          "type": "string"
        },
        "field": {
          "type": "string"
        },
        "keyword": {
          "type": "string"
        },
        "sceneName": {
          "type": "string"
        },
        "userLevel": {
          "type": "string"
        }
      },
      "title": "SceneInfo"
    },
...
    "TbModelInfo": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "isVisible": {
          "type": "boolean"
        },
        "modelName": {
          "type": "string"
        },
        "modelNameZh": {
          "type": "string"
        },
        "serviceType": {
          "type": "string"
        },
        "systemId": {
          "type": "integer",
          "format": "int32"
        },
        "unit": {
          "type": "string"
        },
        "userLevel": {
          "type": "string"
        }
      },
      "title": "TbModelInfo"
    },
...
  }
}